Fix duplicate interceptor CreateProxy helper
Add a regression test that exercises two reflection-activated interceptor signatures for the same inaccessible external proxy. This reproduces the duplicate helper case that caused CS0111/CS0121 in generated Interceptor.g.cs.

Hoist CreateProxy helper emission to the class level in FullModel and emit it once when any interception uses reflection activation, instead of emitting it per interception method.
