+++ src/StreamJsonRpc.Analyzers/GeneratorModels/FullModel.cs
@@ -180,6 +180,26 @@ internal record FullModel
             model.WriteInterceptor(writer);
         }
 
+        if (interceptions.Any(i => i.UseReflectionActivation))
+        {
+            writer.WriteLine("""
+                private static global::StreamJsonRpc.IJsonRpcClientProxy CreateProxy(global::StreamJsonRpc.JsonRpc jsonRpc, in global::StreamJsonRpc.Reflection.ProxyInputs proxyInputs, bool disposeJsonRpcOnFailure)
+                {
+                    if (global::StreamJsonRpc.Reflection.ProxyBase.TryCreateProxy(jsonRpc, proxyInputs, out global::StreamJsonRpc.IJsonRpcClientProxy? proxy))
+                    {
+                        return proxy;
+                    }
+
+                    if (disposeJsonRpcOnFailure)
+                    {
+                        jsonRpc.Dispose();
+                    }
+
+                    throw new global::System.NotImplementedException("Unable to find a source generated proxy filling the specified requirements. Research the NativeAOT topic in the documentation at https://microsoft.github.io/vs-streamjsonrpc");
+                }
+                """);
+        }
+
         writer.WriteLine("""
 
             private static global::StreamJsonRpc.IJsonRpcClientProxy StartListening(global::StreamJsonRpc.IJsonRpcClientProxy proxy)
