+++ test/StreamJsonRpc.Analyzer.Tests/ProxyGeneratorTests.cs
@@ -1149,4 +1149,66 @@ public class ProxyGeneratorTests
 
         await test.RunAsync(TestContext.Current.CancellationToken);
     }
+
+    [Fact]
+    public async Task Interceptor_ExternalProxyInaccessible_DifferentProject_MultipleAttachSignatures()
+    {
+        string libSource = VerifyCS.SourceFilePrefix + /* lang=c#-test */ """
+
+            #nullable enable
+
+            [assembly: ExportRpcContractProxies(ForbidExternalProxyGeneration = true)]
+
+            [JsonRpcContract]
+            [StreamJsonRpc.Reflection.JsonRpcProxyMapping(typeof(StreamJsonRpc.Generated.MyServiceProxy))] // sourcegen runs too late in the test harness, so we have to write it ourselves.
+            public partial interface IMyService
+            {
+            }
+
+            namespace StreamJsonRpc.Generated
+            {
+                internal class MyServiceProxy(JsonRpc client, in StreamJsonRpc.Reflection.ProxyInputs inputs)
+                    : StreamJsonRpc.Reflection.ProxyBase(client, inputs), IMyService
+                {
+                }
+            }
+            """;
+
+        VerifyCS.Test test = new()
+        {
+            TestState =
+            {
+                Sources =
+                {
+                    VerifyCS.SourceFilePrefix + /* lang=c#-test */ """
+
+                    class Test
+                    {
+                        void Foo(System.IO.Stream stream, IJsonRpcMessageHandler handler)
+                        {
+                            IMyService service1 = JsonRpc.Attach<IMyService>(stream);
+                            IMyService service2 = JsonRpc.Attach<IMyService>(handler);
+                        }
+                    }
+                    """,
+                },
+                AdditionalProjects =
+                {
+                    ["ContractsLib"] =
+                    {
+                        Sources =
+                        {
+                            ("IMyService.cs", libSource),
+                        },
+                    },
+                },
+                AdditionalProjectReferences =
+                {
+                    "ContractsLib",
+                },
+            },
+        };
+
+        await test.RunAsync(TestContext.Current.CancellationToken);
+    }
 }
